Change int arguments to double for pwm()
changing inputs to doubles typecasts inputs and allows for correct values

+++ students/Hasan/Lab1/Part2_PWMExploration/pwm_exploration/pwm_exploration.ino
@@ -14,7 +14,7 @@ const int yLED = 9;
  *    none 
  */
 
-void pwm(int pin, int brightness, int period) {
+void pwm(double pin, double brightness, double period) {
   //Calculate Ton
 
   int Ton;
